remove webpack and go all-in on Parcel
