+++ public/index.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <title>Studify</title>
+  <meta charset="UTF-8" />
+  <meta content="width=device-width, initial-scale=1" name="viewport" />
+
+  <!-- markdown highlight -->
+  <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/highlight.js/10.1.1/styles/default.min.css">
+  <script src="//cdnjs.cloudflare.com/ajax/libs/highlight.js/10.1.1/highlight.min.js"></script>
+
+  <link rel="stylesheet" href="./styles.css" />
+</head>
+
+<body>
+  <div id="app"></div>
+  <div id="portal"></div>
+  <!-- WEBPACK -->
+  <!-- <script src="./bundle.js"></script> -->
+  <!-- PARCEL -->
+  <script src="../src/index.jsx"></script>
+</body>
+
+</html>
