--- public/parcel/index.html
@@ -1,27 +1,0 @@
-<!-- THIS IS JUST FOR WEBPACK-DEV-SERVER use 
-since the actual React app is served in the ERBs -->
-
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <title>Studify</title>
-  <meta charset="UTF-8" />
-  <meta content="width=device-width, initial-scale=1" name="viewport" />
-
-  <!-- markdown highlight -->
-  <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/highlight.js/10.1.1/styles/default.min.css">
-  <script src="//cdnjs.cloudflare.com/ajax/libs/highlight.js/10.1.1/highlight.min.js"></script>
-
-  <link rel="stylesheet" href="./styles.css" />
-</head>
-
-<body>
-  <div id="app"></div>
-  <div id="portal"></div>
-  <!-- <script src="./bundle.js"></script> -->
-  <!-- PARCEL -->
-  <script src="../javascripts/index.jsx"></script>
-</body>
-
-</html>
