--- public/parcel/styles.css
@@ -1,47 +1,0 @@
-html,
-body {
-  margin: 0;
-  padding: 0;
-}
-
-body {
-  height: 100vh;
-  width: 100vw;
-
-  font-family: sans-serif;
-  font-size: 20px;
-  color: #000000;
-  background-color: #ffffff;
-}
-
-* {
-  box-sizing: border-box;
-}
-
-h1 {
-  font-size: 2rem;
-}
-
-h2 {
-  margin-top: 0.25rem;
-  margin-bottom: 0.25rem;
-}
-
-#app {
-  height: 100%;
-  width: 100%;
-
-  /* display: flex; */
-  /* flex-direction: column; */
-}
-
-nav ul {
-  list-style-type: none;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-nav ul li {
-  margin-left: 1rem;
-}
