+++ src/Spacer.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+const Spacer = () => {
+  // TODO: pass styles, generalize Field
+  return <div style={{ margin: "0.5rem" }}></div>;
+};
+
+export default Spacer;
